+++ src/backend.rs
@@ -130,6 +130,7 @@ impl LanguageServer for Backend {
                         ">".to_string(),
                         ":".to_string(),
                         "(".to_string(),
+                        "[".to_string(),
                     ]),
                     resolve_provider: Some(true),
                     ..Default::default()